+++ MercadoPagoSDK/MercadoPagoSDK/Core/Checkout/Extensions/MercadoPagoCheckout+PaymentFlowHandler.swift
@@ -12,8 +12,12 @@ extension MercadoPagoCheckout: PXPaymentResultHandlerProtocol {
 
     func finishPaymentFlow(error: MPSDKError) {
         let lastViewController = viewModel.pxNavigationHandler.navigationController.viewControllers.last
-        if let newResultViewController = lastViewController as? PXNewResultViewController {
-            newResultViewController.progressButtonAnimationTimeOut()
+        if lastViewController is PXNewResultViewController || lastViewController is PXSecurityCodeViewController {
+            if let newResultViewController = lastViewController as? PXNewResultViewController {
+                newResultViewController.progressButtonAnimationTimeOut()
+            } else if let securityCodeVC = lastViewController as? PXSecurityCodeViewController {
+                resetButtonAndCleanToken(securityCodeVC: securityCodeVC)
+            }
         }
     }
 
@@ -42,9 +46,14 @@ extension MercadoPagoCheckout: PXPaymentResultHandlerProtocol {
         }
     }
 
+    func resetButtonAndCleanToken(securityCodeVC: PXSecurityCodeViewController) {
+        viewModel.paymentData.cleanToken()
+        securityCodeVC.resetButton()
+    }
+
     private func shouldCallAnimateButton() -> Bool {
         let lastViewController = viewModel.pxNavigationHandler.navigationController.viewControllers.last
-        if lastViewController is PXNewResultViewController {
+        if lastViewController is PXNewResultViewController || lastViewController is PXSecurityCodeViewController {
             return true
         }
         return false