+++ MercadoPagoSDK/MercadoPagoSDK/Services/MercadoPagoServices.swift
@@ -7,6 +7,8 @@
 //
 
 import Foundation
+// Se importa MLCardForm para reutilizar la clase de Reachability
+import MLCardForm
 
 internal class MercadoPagoServices: NSObject {
 
@@ -16,13 +18,19 @@ internal class MercadoPagoServices: NSObject {
     private var branchId: String?
     private var baseURL: String! = PXServicesURLConfigs.MP_API_BASE_URL
     private var gatewayBaseURL: String!
-
+    var reachability: Reachability?
+    var hasInternet: Bool = true
     private var language: String = NSLocale.preferredLanguages[0]
 
     init(publicKey: String, privateKey: String? = nil) {
         self.publicKey = publicKey
         self.privateKey = privateKey
         super.init()
+        addReachabilityObserver()
+    }
+
+    deinit {
+        removeReachabilityObserver()
     }
 
     func update(processingModes: [String]? , branchId: String? = nil) {
@@ -80,7 +88,7 @@ internal class MercadoPagoServices: NSObject {
         service.createPayment(headers: headers, body: paymentDataJSON, params: params, success: callback, failure: failure)
     }
 
-    func getPointsAndDiscounts(url: String, uri: String, paymentIds: [String]? = nil, paymentMethodsIds: [String]? = nil, campaignId: String?, platform: String, ifpe: Bool, headers: [String: String], callback : @escaping (PXPointsAndDiscounts) -> Void, failure: @escaping (() -> Void)) {
+    func getPointsAndDiscounts(url: String, uri: String, paymentIds: [String]? = nil, paymentMethodsIds: [String]? = nil, campaignId: String?, prefId: String?, platform: String, ifpe: Bool, headers: [String: String], callback : @escaping (PXPointsAndDiscounts) -> Void, failure: @escaping (() -> Void)) {
         let service: CustomService = CustomService(baseURL: url, URI: uri)
 
         var params = MercadoPagoServices.getParamsAccessTokenAndPaymentIdsAndPlatform(privateKey, paymentIds, platform)
@@ -88,6 +96,7 @@ internal class MercadoPagoServices: NSObject {
 
         params.paramsAppend(key: ApiParam.API_VERSION, value: PXServicesURLConfigs.API_VERSION)
         params.paramsAppend(key: ApiParam.IFPE, value: String(ifpe))
+        params.paramsAppend(key: ApiParam.PREF_ID, value: prefId)
 
         if let campaignId = campaignId {
             params.paramsAppend(key: ApiParam.CAMPAIGN_ID, value: campaignId)